fix: rename back to RepositoryScheduledTasks
