+++ backend/src/services/index.ts
@@ -2,7 +2,7 @@ import RenderService from './RenderService';
 import EmailService from './email/EmailService';
 import EmailServiceFactory from './email/EmailServiceFactory';
 import ExchangeRateService from './ExchangeRateService';
-import { RepoScheduledTasks } from './scheduled-tasks/ScheduledTaskService';
+import { RepoScheduledTasks } from './scheduled-tasks/RepositoryScheduledTasks';
 import { ServiceScheduledTasks } from './scheduled-tasks/ServiceScheduledTasks';
 import UserAuthService from './UserAuthService';
